+++ DevAppProject/App.js
@@ -7,8 +7,8 @@ import Header from "./src/components/Header/index";
 import Cadastro from './src/screens/Cadastro.js'
 import LoginScreen from "./src/screens/Login.js";
 import VisualizacaoPerfil from "./src/screens/VisualizacaoPerfil.js";
-import CadastroPessoal from "./src/screens/CadastroPessoal.js";
-import EditarPerfil from "./src/screens/EditarPerfil.js";
+import CadastroPessoal from "./src/screens/CadastroPessoal";
+import EditarPerfil from "./src/screens/EditarPerfil";
 import Dashboard from "./src/screens/Dashboard";
 import CadastroAnimal from "./src/screens/CadastroAnimal";
 import { AuthProvider } from "./src/config/auth.js"
@@ -35,23 +35,13 @@ const LoginTela = ({ navigation }) => {
   );
 };
 
-const CadastroPessoalTela = ({ navigation }) => {
-  return (
-    <View style={styles.container}>
-      <Header navigation={navigation} text="Cadastro Pessoal" />
-      <CadastroPessoal navigation={navigation}/>
-      <StatusBar style="auto" />
-    </View>
-  );
-}
-
 export default function App() {
   return (
     <NavigationContainer>
       <AuthProvider>
         <Stack.Navigator screenOptions={{ headerShown: false }}>
           <Stack.Screen name="Cadastro" component={CadastroTela} />
-          <Stack.Screen name="CadastroPessoal" component={CadastroPessoalTela} />
+          <Stack.Screen name="CadastroPessoal" component={CadastroPessoal} />
           <Stack.Screen name="Login" component={LoginTela} />
           <Stack.Screen name="Dashboard" component={Dashboard} />
           <Stack.Screen name="CadastroAnimal" component={CadastroAnimal} />
